requirements.txt
fastapi
uvicorn[standard]
alembic
pytest
python-dotenv
python-jose[cryptography] ## for: jwt token
bcrypt
pydantic[email]